+++ amethyst_audio/src/bundle.rs
@@ -1,13 +1,11 @@
 //! ECS audio bundles
 
-use std::marker::PhantomData;
-
 use amethyst_assets::Processor;
 use amethyst_core::bundle::{Result, SystemBundle};
 use amethyst_core::specs::prelude::DispatcherBuilder;
 use rodio::default_output_device;
-
 use source::*;
+use std::marker::PhantomData;
 use systems::DjSystem;
 
 /// Audio bundle