+++ amethyst_audio/src/components/audio_emitter.rs
@@ -1,13 +1,11 @@
-use std::io::Cursor;
-use std::sync::atomic::AtomicBool;
-use std::sync::Arc;
-
 use amethyst_core::specs::prelude::Component;
 use amethyst_core::specs::storage::BTreeStorage;
 use rodio::{Decoder, SpatialSink};
 use smallvec::SmallVec;
-
 use source::Source;
+use std::io::Cursor;
+use std::sync::atomic::AtomicBool;
+use std::sync::Arc;
 use DecoderError;
 
 /// An audio source, add this component to anything that emits sound.