+++ amethyst_renderer/src/pass/pbm/interleaved.rs
@@ -1,16 +1,13 @@
 //! Forward physically-based drawing pass.
 
-use std::marker::PhantomData;
-
+use super::*;
 use amethyst_assets::AssetStorage;
 use amethyst_core::specs::prelude::{Join, Read, ReadExpect, ReadStorage};
 use amethyst_core::transform::GlobalTransform;
-use gfx::pso::buffer::ElemStride;
-use gfx_core::state::{Blend, ColorMask};
-
-use super::*;
 use cam::{ActiveCamera, Camera};
 use error::Result;
+use gfx::pso::buffer::ElemStride;
+use gfx_core::state::{Blend, ColorMask};
 use light::Light;
 use mesh::{Mesh, MeshHandle};
 use mtl::{Material, MaterialDefaults};
@@ -19,6 +16,7 @@ use pass::util::{draw_mesh, get_camera, setup_textures, setup_vertex_args};
 use pipe::pass::{Pass, PassData};
 use pipe::{DepthMode, Effect, NewEffect};
 use resources::AmbientColor;
+use std::marker::PhantomData;
 use tex::Texture;
 use types::{Encoder, Factory};
 use vertex::{Normal, Position, Query, Tangent, TexCoord};