+++ amethyst_renderer/src/resources.rs
@@ -2,10 +2,9 @@
 
 use amethyst_assets::{PrefabData, PrefabError};
 use amethyst_core::specs::{Entity, Write};
+use color::Rgba;
 use smallvec::SmallVec;
 use winit::Window;
-
-use color::Rgba;
 
 /// The ambient color of a scene
 #[derive(Clone, Debug, Default, Serialize, Deserialize)]